requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
ffmpeg-python==0.2.0
openai-whisper==20231117
numpy==1.26.4
pandas==2.2.2
librosa==0.10.2.post1
soundfile==0.12.1
matplotlib==3.8.4
transformers==4.41.2
torch>=2.2.0
textblob==0.18.0.post0
nltk==3.8.1
keybert==0.8.5
yake==0.4.8
scikit-learn==1.5.1
mediapipe==0.10.14
opencv-python==4.10.0.84
youtube-transcript-api==0.6.2
google-api-python-client==2.140.0
requests==2.32.3